+++ packages/babel-plugin-pob-babel/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.8](https://github.com/christophehurpeau/pob/compare/babel-plugin-pob-babel@0.7.7...babel-plugin-pob-babel@0.7.8) (2019-03-21)
+
+**Note:** Version bump only for package babel-plugin-pob-babel
+
+
+
+
+
 ## [0.7.7](https://github.com/christophehurpeau/pob/compare/babel-plugin-pob-babel@0.7.6...babel-plugin-pob-babel@0.7.7) (2019-03-09)
 
 **Note:** Version bump only for package babel-plugin-pob-babel