+++ packages/babel-plugin-transform-builtins/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/christophehurpeau/pob/compare/babel-plugin-transform-builtins@1.0.0...babel-plugin-transform-builtins@1.0.1) (2019-03-21)
+
+**Note:** Version bump only for package babel-plugin-transform-builtins
+
+
+
+
+
 # 1.0.0 (2019-03-09)