+++ packages/pob-lcov-reporter/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/christophehurpeau/pob/compare/pob-lcov-reporter@1.0.1...pob-lcov-reporter@1.0.2) (2019-03-21)
+
+**Note:** Version bump only for package pob-lcov-reporter
+
+
+
+
+
 ## [1.0.1](https://github.com/christophehurpeau/pob/compare/pob-lcov-reporter@1.0.0...pob-lcov-reporter@1.0.1) (2019-03-09)
 
 **Note:** Version bump only for package pob-lcov-reporter